requirements.txt
black==23.11.0
click==8.1.7
mypy-extensions==1.0.0
numpy==1.26.2
numpy-quaternion==2022.4.3
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
tomli==2.0.1
typing_extensions==4.8.0
